requirements.txt
python-telegram-bot==20.7
pdfplumber==0.10.3
gTTS==2.4.0
python-dotenv==1.0.0